[rev] replace CrosshairViewModel with the generalized vm originally implemented in CrosshairWorldViewModel
