+++ PresentationTest/App.xaml.cs
@@ -13,6 +13,6 @@ public partial class App : PrismApplication
     }
 
     protected override Window CreateShell() =>
-         // Container.Resolve<VtkMvvmTestWindow>();
-        Container.Resolve<VtkObliqueSliceTestWindow>();
+         Container.Resolve<VtkMvvmTestWindow>();
+        // Container.Resolve<VtkObliqueSliceTestWindow>();
 }
